Week 16/Memory Agent/requirements.txt
# Memory-Controlled Conversational Agent Requirements

# Core Dependencies
langgraph
langchain-core
langchain-openai


langgraph-cli

python-dotenv
